+++ Backend/.gitignore
@@ -1,6 +1,26 @@
-__pycache__
-ngrok.exe
-universe_server.pem
-venv
-texto.txt
+# Python compiled files
+*.pyc
+*.pyo
+__pycache__/
+
+# Environment variables and virtual environments
+.venv/
 .env
+.env.dev
+.env.production
+.env.staging
+*.env
+venv/
+env/
+.virtualenv/
+
+# Logs and debug files
+*.log
+*.sqlite3
+
+# IDE and editor specific files
+.vscode/
+.idea/
+
+# Ignore specific backend files or folders
+/backend_folder_name_to_ignore/
